+++ R/tools.R
@@ -1,9 +1,11 @@
 # internal tools
 check_lm <- function(model) stopifnot(inherits(model, "lm"))
+check_lm_forced <- function(model) stopifnot(inherits(model, c("lm", "lm_forced")))
 
 values_lm <- function(model, type = c("auto", "linear", "power")){
   type <- match.arg(type)
-  check_lm(model)
+  # check_lm(model)
+  check_lm_forced(model)
 
   if(type == "auto"){
     if(check_power(model)) type <- "power"
@@ -11,14 +13,25 @@ values_lm <- function(model, type = c("auto", "linear", "power")){
   }
 
   ans <- NULL
-
-  ans$p <- model$rank
-  ans$rdf <- model$df.residual
-  ans$r <- model$residuals
-  ans$f <- model$fitted.values
-  ans$y <- stats::model.frame(model)[[1]]
-  ans$n <- length(ans$y)
-  ans$k <- length(stats::coef(model))
+  if(inherits(model, "lm")){
+    ans$p <- model$rank
+    ans$rdf <- model$df.residual
+    ans$r <- model$residuals
+    ans$f <- model$fitted.values
+    ans$y <- stats::model.frame(model)[[1]]
+    ans$n <- length(ans$y)
+    ans$k <- length(stats::coef(model))
+    ans$df_int <- attr(model$terms, "intercept")
+  }else if(inherits(model, "lm_forced")){
+    ans$p   <- model$rank
+    ans$rdf <- model$df.residual
+    ans$r   <- as.vector(model$residuals)
+    ans$f   <- as.vector(model$fitted_values)
+    ans$y   <- as.vector(model$y)
+    ans$n   <- length(ans$y)
+    ans$k   <- length(model$Estimate)
+    ans$df_int <- ifelse(model$intercept, 1, 0)
+  }
 
   if(type == "power"){
     ans$r <- exp(ans$r)
@@ -28,7 +41,6 @@ values_lm <- function(model, type = c("auto", "linear", "power")){
 
   ans$e <- ans$y - ans$f
 
-  ans$df_int <- attr(model$terms, "intercept")
   ans$a <- (ans$n - ans$df_int) / ans$rdf
 
   ans$type <- type
@@ -42,12 +54,19 @@ check_power <- function(model){
 }
 
 lm_forced_int <- function(model){
-  check_lm(model)
+  check_lm_forced(model)
 
-  mf <- stats::model.frame(model)
-  if(attr(model$terms, "intercept")) X <- stats::model.matrix(stats::formula(model), mf)
-  else X <- cbind(`Intercept` = 1, stats::model.matrix(stats::formula(model), mf))
-  y <- stats::model.response(mf)
+  if(inherits(model, "lm")){
+    mf <- stats::model.frame(model)
+    if(attr(model$terms, "intercept")) X <- stats::model.matrix(stats::formula(model), mf)
+    else X <- cbind(`(Intercept)` = 1, stats::model.matrix(stats::formula(model), mf))
+    y <- stats::model.response(mf)
+  }else if(inherits(model, "lm_forced")){
+    mf <- model$mf
+    if(model$intercept) X <- model$X
+    else X <- cbind(`(Intercept)` = 1, model$X)
+    y <- model$y
+  }
 
   qr_decomp <- qr(X)
 
@@ -71,13 +90,77 @@ lm_forced_int <- function(model){
   weigths <- stats::weights(model)
 
   results <- list(
+    mf = mf,
+    X = X,
     Estimate = betas,
     Std.Error = std_errors,
+    df.residual = df_residual,
+    rank = ncol(X),
     t_value = betas / std_errors,
+    y = y,
     fitted_values = fitted_values,
     residuals = residuals,
-    weights = weigths
+    weights = weigths,
+    intercept = TRUE
   )
+
+  class(results) <- "lm_forced"
+
+  results
+}
+
+lm_forced_without_int <- function(model){
+  check_lm_forced(model)
+
+  if(inherits(model, "lm")){
+    mf <- stats::model.frame(model)
+    if(!attr(model$terms, "intercept")) X <- stats::model.matrix(stats::formula(model), mf)
+    else X <- stats::model.matrix(stats::formula(model), mf)[, -1, drop = FALSE]
+    y <- stats::model.response(mf)
+  }else if(inherits(model, "lm_forced")){
+    mf <- model$mf
+    if(!model$intercept) X <- model$X
+    else X <- model$X[, -1, drop = FALSE]
+    y <- model$y
+  }
+
+  qr_decomp <- qr(X)
+
+  betas <- qr.coef(qr_decomp, y)
+
+  residuals <- y - X %*% betas
+  df_residual <- nrow(X) - ncol(X)
+
+  sigma_sq <- sum(residuals^2) / df_residual
+
+  R <- qr.R(qr_decomp)
+  v_cov <- solve(t(R) %*% R) * sigma_sq
+  std_errors <- sqrt(diag(v_cov))
+
+  # betas_vec <- as.matrix(betas)
+  # fitted_values <- X %*% betas_vec
+
+  Q <- qr.Q(qr_decomp)
+  fitted_values <- Q %*% (t(Q) %*% y)
+
+  weigths <- stats::weights(model)
+
+  results <- list(
+    mf = mf,
+    X = X,
+    Estimate = betas,
+    Std.Error = std_errors,
+    df.residual = df_residual,
+    rank = ncol(X),
+    t_value = betas / std_errors,
+    y = y,
+    fitted_values = fitted_values,
+    residuals = residuals,
+    weights = weigths,
+    intercept = FALSE
+  )
+
+  class(results) <- "lm_forced"
 
   results
 }